Migrate test project location.
