--- test/Ocelot.Discovery.Nacos.UnitTest/NacosProviderFactoryTests.cs
@@ -1,146 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Moq;
-using Ocelot.Discovery.Nacos;
-using Ocelot.Configuration;
-using Ocelot.ServiceDiscovery;
-using System;
-using Nacos.V2;
-
-namespace Ocelot.Tests.Discovery.Nacos
-{
-    [TestClass]
-    public class NacosProviderFactoryTests
-    {
-        [TestMethod]
-        public void CreateProvider_ShouldReturnNacosProvider_WhenConfigTypeIsNacos()
-        {
-            // Arrange
-            var serviceProviderMock = new Mock<IServiceProvider>();
-            var nacosNamingServiceMock = new Mock<INacosNamingService>();
-            serviceProviderMock.Setup(sp => sp.GetService(typeof(INacosNamingService)))
-                               .Returns(nacosNamingServiceMock.Object);
-
-            var config = new ServiceProviderConfiguration(
-                type: "nacos",
-                scheme: "http",
-                host: "localhost",
-                port: 8848,
-                token: null,
-                configurationKey: "nacos",
-                pollingInterval: 5000
-            );
-
-            var route = new DownstreamRoute(
-                key: "testKey",
-                upstreamPathTemplate: null,
-                upstreamHeadersFindAndReplace: null,
-                downstreamHeadersFindAndReplace: null,
-                downstreamAddresses: null,
-                serviceName: "testService",
-                serviceNamespace: null,
-                httpHandlerOptions: null,
-                useServiceDiscovery: false,
-                enableEndpointEndpointRateLimiting: false,
-                qosOptions: null,
-                downstreamScheme: null,
-                requestIdKey: null,
-                isCached: false,
-                cacheOptions: null,
-                loadBalancerOptions: null,
-                rateLimitOptions: null,
-                routeClaimsRequirement: null,
-                claimsToQueries: null,
-                claimsToHeaders: null,
-                claimsToClaims: null,
-                claimsToPath: null,
-                isAuthenticated: false,
-                isAuthorized: false,
-                authenticationOptions: null,
-                downstreamPathTemplate: null,
-                loadBalancerKey: null,
-                delegatingHandlers: null,
-                addHeadersToDownstream: null,
-                addHeadersToUpstream: null,
-                dangerousAcceptAnyServerCertificateValidator: false,
-                securityOptions: null,
-                downstreamHttpMethod: null,
-                downstreamHttpVersion: null,
-                downstreamHttpVersionPolicy: default,
-                upstreamHeaders: null,
-                metadataOptions: null
-            );
-
-            // Act
-            var provider = NacosProviderFactory.Get(serviceProviderMock.Object, config, route);
-
-            // Assert
-            Assert.IsNotNull(provider);
-            Assert.IsInstanceOfType(provider, typeof(Ocelot.Discovery.Nacos.Nacos));
-        }
-
-        [TestMethod]
-        [ExpectedException(typeof(NullReferenceException))]
-       public void CreateProvider_ShouldThrowException_WhenNacosNamingServiceIsNull()
-        {
-            // Arrange
-            var serviceProviderMock = new Mock<IServiceProvider>();
-            serviceProviderMock.Setup(sp => sp.GetService(typeof(INacosNamingService)))
-                               .Returns(null);
-
-            var config = new ServiceProviderConfiguration(
-                type: "nacos",
-                scheme: "http",
-                host: "localhost",
-                port: 8848,
-                token: null,
-                configurationKey: "nacos",
-                pollingInterval: 5000
-            );
-
-            var route = new DownstreamRoute(
-                key: "testKey",
-                upstreamPathTemplate: null,
-                upstreamHeadersFindAndReplace: null,
-                downstreamHeadersFindAndReplace: null,
-                downstreamAddresses: null,
-                serviceName: "testService",
-                serviceNamespace: null,
-                httpHandlerOptions: null,
-                useServiceDiscovery: false,
-                enableEndpointEndpointRateLimiting: false,
-                qosOptions: null,
-                downstreamScheme: null,
-                requestIdKey: null,
-                isCached: false,
-                cacheOptions: null,
-                loadBalancerOptions: null,
-                rateLimitOptions: null,
-                routeClaimsRequirement: null,
-                claimsToQueries: null,
-                claimsToHeaders: null,
-                claimsToClaims: null,
-                claimsToPath: null,
-                isAuthenticated: false,
-                isAuthorized: false,
-                authenticationOptions: null,
-                downstreamPathTemplate: null,
-                loadBalancerKey: null,
-                delegatingHandlers: null,
-                addHeadersToDownstream: null,
-                addHeadersToUpstream: null,
-                dangerousAcceptAnyServerCertificateValidator: false,
-                securityOptions: null,
-                downstreamHttpMethod: null,
-                downstreamHttpVersion: null,
-                downstreamHttpVersionPolicy: default,
-                upstreamHeaders: null,
-                metadataOptions: null
-            );
-
-            // Act
-            var provider = NacosProviderFactory.Get(serviceProviderMock.Object, config, route);
-            // Assert
-            Assert.IsNull(provider);
-        }
-    }
-}
