+++ unit/Ocelot.Discovery.Nacos.UnitTest.csproj
@@ -18,7 +18,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\src\Ocelot.Discovery.Nacos\Ocelot.Discovery.Nacos.csproj" />
+      <ProjectReference Include="..\src\Ocelot.Discovery.Nacos\Ocelot.Discovery.Nacos.csproj" />
     </ItemGroup>
 
 </Project>
